move common schema to common schema file and move THAT file to a shared location
