+++ x-pack/plugins/files/server/routes/file_kind/download.ts
@@ -10,8 +10,9 @@ import { Ensure } from '@kbn/utility-types';
 import { Readable } from 'stream';
 
 import type { DownloadFileKindHttpEndpoint } from '../../../common/api_routes';
+import { fileNameWithExt } from '../common_schemas';
 import { fileErrors } from '../../file';
-import { getDownloadHeadersForFile, fileNameSchema } from '../common';
+import { getDownloadHeadersForFile } from '../common';
 import { getById } from './helpers';
 import type { FileKindsRequestHandler } from './types';
 
@@ -19,7 +20,7 @@ export const method = 'get' as const;
 
 export const paramsSchema = schema.object({
   id: schema.string(),
-  fileName: schema.maybe(fileNameSchema),
+  fileName: schema.maybe(fileNameWithExt),
 });
 
 type Params = Ensure<DownloadFileKindHttpEndpoint['inputs']['params'], TypeOf<typeof paramsSchema>>;
